requirements.txt
telethon>=1.29.0
python-dotenv>=1.0.0